fix: proper type for route indices
The value type should be `uint32_t` to be consistent with the LPM value
type and other functions in the "route" module.

+++ dataplane/unittest/main.c
@@ -139,7 +139,7 @@ main(int argc, char **argv) {
 		}
 	);
 
-	route_module_config_add_route_list(rmc, 1, (uint64_t[]){0});
+	route_module_config_add_route_list(rmc, 1, (uint32_t[]){0});
 
 	route_module_config_add_prefix_v4(
 		rmc,
